add error type for database/decodeString and refactor error handling for root repository
