+++ be1-go/internal/sqlite/sqlite.go
@@ -479,82 +479,6 @@ func (s *SQLite) HasMessage(messageID string) (bool, error) {
 // RootRepository interface implementation
 //======================================================================================================================
 
-func (s *SQLite) StoreLaoWithLaoGreet(
-	channels map[string]string,
-	laoPath string,
-	organizerPubBuf []byte,
-	msg, laoGreetMsg message.Message) error {
-
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	tx, err := s.database.Begin()
-	if err != nil {
-		return err
-	}
-
-	msgByte, err := json.Marshal(msg)
-	if err != nil {
-		return err
-	}
-	laoGreetMsgByte, err := json.Marshal(laoGreetMsg)
-	if err != nil {
-		return err
-	}
-
-	messageData, err := base64.URLEncoding.DecodeString(msg.Data)
-	if err != nil {
-		return err
-	}
-	laoGreetData, err := base64.URLEncoding.DecodeString(laoGreetMsg.Data)
-	if err != nil {
-		return err
-	}
-
-	storedTime := time.Now().UnixNano()
-
-	for channel, channelType := range channels {
-		_, err = tx.Exec(insertChannel, channel, channelTypeToID[channelType], laoPath)
-		if err != nil {
-			return err
-		}
-	}
-
-	err = s.insertMessageHelper(tx, msg.MessageID, msgByte, messageData, storedTime)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertChannelMessage, "/root", msg.MessageID, true)
-	if err != nil {
-		return err
-	}
-
-	_, err = tx.Exec(insertChannelMessage, laoPath, msg.MessageID, false)
-	if err != nil {
-		return err
-	}
-
-	_, err = tx.Exec(insertPublicKey, laoPath, organizerPubBuf)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertMessage, laoGreetMsg.MessageID, laoGreetMsgByte, laoGreetData, storedTime)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertChannelMessage, laoPath, laoGreetMsg.MessageID, false)
-	if err != nil {
-		return err
-	}
-
-	err = tx.Commit()
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
 //======================================================================================================================
 // LaoRepository interface implementation
 //======================================================================================================================
@@ -574,273 +498,6 @@ func (s *SQLite) GetOrganizerPubKey(laoPath string) (kyber.Point, error) {
 		return nil, err
 	}
 	return organizerPubKey, nil
-}
-
-func (s *SQLite) GetRollCallState(channelPath string) (string, error) {
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	var state string
-	err := s.database.QueryRow(selectLastRollCallMessage, messagedata.RollCallObject, channelPath).Scan(&state)
-	if err != nil {
-		return "", err
-	}
-	return state, nil
-}
-
-func (s *SQLite) CheckPrevOpenOrReopenID(channel, nextID string) (bool, error) {
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	var lastMsg []byte
-	var lastAction string
-
-	err := s.database.QueryRow(selectLastRollCallMessageInList, channel, messagedata.RollCallObject,
-		messagedata.RollCallActionOpen, messagedata.RollCallActionReOpen).Scan(&lastMsg, &lastAction)
-
-	if err != nil && errors.Is(err, sql.ErrNoRows) {
-		return false, nil
-	} else if err != nil {
-		return false, err
-	}
-
-	switch lastAction {
-	case messagedata.RollCallActionOpen:
-		var rollCallOpen messagedata.RollCallOpen
-		err = json.Unmarshal(lastMsg, &rollCallOpen)
-		if err != nil {
-			return false, err
-		}
-		return rollCallOpen.UpdateID == nextID, nil
-	case messagedata.RollCallActionReOpen:
-		var rollCallReOpen messagedata.RollCallReOpen
-		err = json.Unmarshal(lastMsg, &rollCallReOpen)
-		if err != nil {
-			return false, err
-		}
-		return rollCallReOpen.UpdateID == nextID, nil
-	}
-
-	return false, nil
-}
-
-func (s *SQLite) CheckPrevCreateOrCloseID(channel, nextID string) (bool, error) {
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	var lastMsg []byte
-	var lastAction string
-
-	err := s.database.QueryRow(selectLastRollCallMessageInList, channel, messagedata.RollCallObject,
-		messagedata.RollCallActionCreate, messagedata.RollCallActionClose).Scan(&lastMsg, &lastAction)
-
-	if err != nil && errors.Is(err, sql.ErrNoRows) {
-		return false, nil
-	} else if err != nil {
-		return false, err
-	}
-
-	switch lastAction {
-	case messagedata.RollCallActionCreate:
-		var rollCallCreate messagedata.RollCallCreate
-		err = json.Unmarshal(lastMsg, &rollCallCreate)
-		if err != nil {
-			return false, err
-		}
-		return rollCallCreate.ID == nextID, nil
-	case messagedata.RollCallActionClose:
-		var rollCallClose messagedata.RollCallClose
-		err = json.Unmarshal(lastMsg, &rollCallClose)
-		if err != nil {
-			return false, err
-		}
-		return rollCallClose.UpdateID == nextID, nil
-	}
-
-	return false, nil
-}
-
-func (s *SQLite) GetLaoWitnesses(laoPath string) (map[string]struct{}, error) {
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	var witnesses []string
-	err := s.database.QueryRow(selectLaoWitnesses, laoPath, messagedata.LAOObject, messagedata.LAOActionCreate).Scan(&witnesses)
-	if err != nil {
-		return nil, err
-	}
-
-	var witnessesMap = make(map[string]struct{})
-	for _, witness := range witnesses {
-		witnessesMap[witness] = struct{}{}
-	}
-
-	return witnessesMap, nil
-}
-
-func (s *SQLite) StoreRollCallClose(channels []string, laoPath string, msg message.Message) error {
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	tx, err := s.database.Begin()
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback()
-
-	messageData, err := base64.URLEncoding.DecodeString(msg.Data)
-	if err != nil {
-		return err
-	}
-
-	msgBytes, err := json.Marshal(msg)
-	if err != nil {
-		return err
-	}
-
-	err = s.insertMessageHelper(tx, msg.MessageID, msgBytes, messageData, time.Now().UnixNano())
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertChannelMessage, laoPath, msg.MessageID, true)
-	if err != nil {
-		return err
-	}
-
-	if len(channels) == 0 {
-		return tx.Commit()
-	}
-
-	for _, channel := range channels {
-		_, err = tx.Exec(insertChannel, channel, channelTypeToID[ChirpType], laoPath)
-		if err != nil {
-			return err
-		}
-	}
-	err = tx.Commit()
-	if err != nil {
-		return err
-	}
-	return nil
-
-}
-
-func (s *SQLite) storeElectionHelper(
-	tx *sql.Tx,
-	storedTime int64,
-	laoPath, electionPath string,
-	electionPubKey kyber.Point,
-	electionSecretKey kyber.Scalar,
-	msg message.Message) error {
-
-	msgBytes, err := json.Marshal(msg)
-	if err != nil {
-		return err
-	}
-	messageData, err := base64.URLEncoding.DecodeString(msg.Data)
-	if err != nil {
-		return err
-	}
-
-	electionPubBuf, err := electionPubKey.MarshalBinary()
-	if err != nil {
-		return err
-	}
-	electionSecretBuf, err := electionSecretKey.MarshalBinary()
-	if err != nil {
-		return err
-	}
-
-	err = s.insertMessageHelper(tx, msg.MessageID, msgBytes, messageData, storedTime)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertChannelMessage, laoPath, msg.MessageID, true)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertChannel, electionPath, channelTypeToID[ElectionType], laoPath)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertChannelMessage, electionPath, msg.MessageID, false)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertKeys, electionPath, electionPubBuf, electionSecretBuf)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (s *SQLite) StoreElection(
-	laoPath, electionPath string,
-	electionPubKey kyber.Point,
-	electionSecretKey kyber.Scalar,
-	msg message.Message) error {
-
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	tx, err := s.database.Begin()
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback()
-
-	storedTime := time.Now().UnixNano()
-
-	err = s.storeElectionHelper(tx, storedTime, laoPath, electionPath, electionPubKey, electionSecretKey, msg)
-	if err != nil {
-		return err
-	}
-
-	return tx.Commit()
-}
-
-func (s *SQLite) StoreElectionWithElectionKey(
-	laoPath, electionPath string,
-	electionPubKey kyber.Point,
-	electionSecretKey kyber.Scalar,
-	msg, electionKeyMsg message.Message) error {
-
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	tx, err := s.database.Begin()
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback()
-
-	storedTime := time.Now().UnixNano()
-
-	err = s.storeElectionHelper(tx, storedTime, laoPath, electionPath, electionPubKey, electionSecretKey, msg)
-	if err != nil {
-		return err
-	}
-
-	electionKey, err := base64.URLEncoding.DecodeString(electionKeyMsg.Data)
-	if err != nil {
-		return err
-	}
-	electionKeyMsgBytes, err := json.Marshal(electionKeyMsg)
-	if err != nil {
-		return err
-	}
-
-	_, err = tx.Exec(insertMessage, electionKeyMsg.MessageID, electionKeyMsgBytes, electionKey, storedTime)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertChannelMessage, electionPath, electionKeyMsg.MessageID, false)
-	if err != nil {
-		return err
-	}
-
-	return tx.Commit()
 }
 
 //======================================================================================================================
